usuarios/views.py
from rest_framework import viewsets
from .models import Usuario, Habilidade
from .serializers import UsuarioSerializer, HabilidadeSerializer

class UsuarioViewSet(viewsets.ModelViewSet):
    queryset = Usuario.objects.all()
    serializer_class = UsuarioSerializer

class HabilidadeViewSet(viewsets.ModelViewSet):
    queryset = Habilidade.objects.all()
    serializer_class = HabilidadeSerializer
